+++ changelog.rst
@@ -5,6 +5,7 @@ dev
 - change Doctrine class name to Assertion
 - change Predicate param name from template to content
 - add Marshmallow schemas
+- add APISpec for documenting Marshmallow schemas
 
 0.4.0 (2021-04-08)
 ------------------